add git clone. For now until I find a way to initialize the folder securly content of an existing app isn't deleted.
It works like it :

git clone appuri [appdir]

or appuri is uri of _design document, for example:

couchapp clone http://localhost:5984/testing4/_design%2Ftest

If a manifest has been created via push, It will use push to recreate
all the app architecture on the disk. If not it will try to recreate it.

Where the is a manigest, there isn't yet any detection of changes in the
doc / manifest. This is on my list.