New managements of dbs from configuration files related to the discussion on ml.
Now a configuration file could be like this one :
{
    "env": {
        "default": {
            "db": "http://127.0.0.1:5984/testing8"
        },
        "dev": {
            "db": [ "http://127.0.0.1:5984/testing9",
"http://127.0.0.1:5984/testing10" ]
        }
    }
}

default db is in env.default property,

you could setup other environnement like dev. a db key could be a string
oor a list. A list allow you to push on different db in one pass.

So now push command is like this :

in a couchapp folder :
`couch push`will use default deb to send the data
`couch push dbstring' will send to dbstring env or if no env with this
name is found it will send to http://127.0.0.1:5984/dbstring
`couch push appname dbstring`

above acouchappfolder command is :

ccouchapp push folder [appname] [dbstring]
dbstring could be a full uri to the db or an environment key or just
name of db (in this case it will send to
http://127.0.0.1:5984/dbstring). If it's not provided it will try to
send to default deb if it has been defined